### 2024-08-17 #7114 @editorialbot
[PRE REVIEW]: Bayesian model reconstruction based on spectral line observations with pomme
**Submitting author:** @FredDeCeuster (<a href="http://orcid.org/0000-0001-5887-8498">Frederik De Ceuster</a>)
**Repository:** https://github.com/Magritte-code/pomme
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.16
