+++ tests/NLTE.py
@@ -1,0 +1,236 @@
+import healpy as hp
+import torch
+import math
+import gc
+import numpy as np
+import matplotlib.pyplot as plt
+from astropy        import units, constants
+import astropy.constants as constants
+from p3droslo.plot  import plot_cube_2D, plot_spectrum
+from p3droslo.model import TensorModel
+from p3droslo.lines import Line
+from p3droslo.haar  import Haar
+import copy
+import p3droslo
+print(p3droslo.__version__)
+from astroquery.lamda import Lamda
+
+CC  = constants.c  .si.value   # Speed of light       [m/s]
+HH  = constants.h  .si.value   # Planck's constant    [J s]
+KB  = constants.k_B.si.value   # Boltzmann's constant [J/K]
+AMU = constants.u  .si.value   # Atomic mass unit     [kg]
+
+
+def compute_sigma(line,temperature,v_turbulence):
+
+    factor_1 = line.frequency / CC
+    factor_2 = 2.0 * KB / (line.species_molar_mass * AMU)
+    return factor_1 * torch.sqrt(factor_2*temperature + v_turbulence**2)
+def compute_profile(frequencies,linefreq,sigma):
+    #print(frequencies.shape)
+    profile=torch.zeros(*sigma.shape,len(frequencies))
+    #while frequencies.dim() < linefreq.dim()+1:
+    #    frequencies = frequencies.unsqueeze(0)
+    mask=sigma>1e-20
+    profile[mask]=(1/(np.sqrt(np.pi)*sigma[...,None][mask]))*torch.exp(-(frequencies[None,None,None,...]-linefreq[...,None][mask])**2/(sigma[...,None][mask]**2))
+    #profile=(1/(torch.pi*sigma[...,None]))*torch.exp(-(frequencies[None,None,None,:]-linefreq)**2/(sigma[...,None]**2))
+    return profile.float()
+def source_ij(line,pop):
+    """
+    Line source function
+
+    Parameters
+    ----------
+    pop : torch.Tensor
+        level populations
+    
+    Returns
+    -------
+    src : torch.Tensor
+    Tensor containing the line source function (= emissivity / opacity).
+    """
+    pop=pop.to(torch.float32)
+    upper_mask=pop[line.upper,:]>1e-20 
+    lower_mask=pop[line.lower,:]>1e-20
+    mask=upper_mask&lower_mask
+    source=torch.zeros(mask.shape)
+    source[mask]=(line.Einstein_A  * pop[line.upper][mask] / (line.Einstein_Ba * pop[line.lower][mask] - line.Einstein_Bs * pop[line.upper][mask]) )
+    return source
+def emissivity_and_opacity_ij(line,model, pop):
+    """
+    Line opacity, not folded with the profile.
+        
+    Parameters
+    ----------
+    pop : torch.Tensor
+        level populations
+    
+    Returns
+    -------
+    eta, chi : torch.Tensor, torch.Tensor
+    Tensor containing the LTE emissivities and opacities for the given temperature.
+    """
+    upper_mask=pop[line.upper,:]>1e-30 
+    lower_mask=pop[line.lower,:]>1e-30
+    mask=upper_mask&lower_mask
+    eta=torch.zeros(model.shape)
+    chi=torch.zeros(model.shape)
+    # Compute the prefactor
+    factor = HH * line.frequency / (4.0 * np.pi)
+        
+    # Compute the emissivity and opacity
+    eta[mask] = factor *  line.Einstein_A  * pop[line.upper][mask]
[…]
