+++ tests/RotatingModel.py
@@ -1,0 +1,127 @@
+import healpy as hp
+import torch
+import math
+import gc
+import numpy as np
+import matplotlib.pyplot as plt
+from astropy        import units, constants
+import astropy.constants as constants
+from p3droslo.plot  import plot_cube_2D, plot_spectrum
+from p3droslo.model import TensorModel
+from p3droslo.lines import Line
+from p3droslo.haar  import Haar
+import copy
+#check whether the version is correct
+#the latest version should be 0.0.15, at time of writing
+import p3droslo
+print(p3droslo.__version__)
+from astroquery.lamda import Lamda
+CC  = constants.c  .si.value   # Speed of light       [m/s]
+HH  = constants.h  .si.value   # Planck's constant    [J s]
+KB  = constants.k_B.si.value   # Boltzmann's constant [J/K]
+AMU = constants.u  .si.value   # Atomic mass unit     [kg]
+
+class RotationModel:
+    def __init__(self, model,pops ,theta, phi,padding=False):
+        self.model= copy.deepcopy(model) 
+        self.theta=theta
+        self.phi=phi
+        self.padding=padding
+        
+        if self.padding is True:
+            self.model=self.pad_model(model)
+        self.rotation_model = copy.deepcopy(self.model)  
+        R,center=self.rotation_matrix(theta,phi,model)
+        self.R=R
+        self.center=center
+        for key in model.keys():
+            if key.startswith('velocity'):
+                pass 
+            else:
+                self.rotation_model[key]=(self.rotate_scalar(self.model[key],center,R)[0])
+                self.mask_region=(self.rotate_scalar(self.model[key],center,R)[1])
+        if 'velocity_x:v_max' in self.rotation_model.keys() and 'velocity_y:v_max' in self.rotation_model.keys() and 'velocity_z:v_max' in self.rotation_model.keys():
+            vx=self.rotate_scalar(model['velocity_x:v_max'],center,R)[0]
+            vy=self.rotate_scalar(model['velocity_y:v_max'],center,R)[0]
+            vz=self.rotate_scalar(model['velocity_z:v_max'],center,R)[0]
+            self.rotation_model['velocity_x:v_max'], self.rotation_model['velocity_y:v_max'],self.rotation_model['velocity_z:v_max']=self.rotate_vector(vx, vy, vz, R)
+        if 'velocity_los' in self.rotation_model.keys():
+            self.rotation_model['velocity_los']=self.rotation_model['velocity_z:v_max']
+        rotated_pops_list = []
+        for i in range(pops.shape[0]):
+            rotated_pop = self.rotate_scalar(pops[i,...], center, R)[0]
+            rotated_pops_list.append(rotated_pop)
+        self.pops = torch.stack(rotated_pops_list, dim=0)
+    def __getitem__(self, key):
+        # This method allows subscript notation.
+        return self.rotation_model[key]
+    def pad_model(self,model):
+        dz=model.dx(2)
+        diagonal_length = int(torch.ceil(torch.norm(torch.tensor(model.shape, dtype=torch.float)).item()))
+        new_sizes=diagonal_length*dz
+        padded_model= TensorModel(shape=(diagonal_length,diagonal_length,diagonal_length), sizes=3*(new_sizes,), dtau_warning_threshold=0.1)
+        pad_dim1 = int(torch.ceil((diagonal_length - model.shape[0])/2))
+        pad_dim2 = int(torch.ceil((diagonal_length - model.shape[1])/2))
+        pad_dim3 = int(torch.ceil((diagonal_length - model.shape[2])/2))
[…]
