### 2024-10-07 #7114 @JBorrow
Hi @christopherlovell, would you be available to review this package for JOSS? The reviewer's guide is here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
