### 2024-10-07 #7114 @karllark
> Hi @karllark, I found your name in the JOSS reviewer's database and think you'd be a great fit as a reviewer of this paper. Would that be something you'd be interested in? Our reviewer's guide is available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Yes, I'd be happy to review this paper.
