### 2024-10-07 #7321 @editorialbot
[REVIEW]: Bayesian model reconstruction based on spectral line observations with pomme
**Submitting author:** @FredDeCeuster (<a href="http://orcid.org/0000-0001-5887-8498">Frederik De Ceuster</a>)
**Repository:** https://github.com/Magritte-code/pomme
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.0
