### 2024-10-10 #7114 @JBorrow
Hi @s-goldman, I found you as a corresponding author to a similar submission to JOSS as this one and think you'd be a great reviewer for this paper. Would that be something you'd be interested in? Our reviewer's guide is available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
