### 2024-10-10 #7321 @karllark
I will get started on my review after the JWST deadline and probably the STScI Flux Calibration workshop.  Hence very late Oct or early Nov.  Will try to get it before then, but not sure I will be able to.  Just a FYI.
