### 2024-10-14 #7114 @christopherlovell
> Hi @christopherlovell, would you be available to review this package for JOSS? The reviewer's guide is here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Yes can review, but not until early November
