### 2024-10-22 Magritte-code/pomme#8 @s-goldman
[JOSS Review]: Paper comments 
I've found a couple things that you may want to include in the paper. Not necessarily blockers.

1. The target demographic for this code is not explicitly mentioned anywhere. Some detail may allow a user to determine the level of expertise required/expected.
2. Related to that point, "optically thin" may need more explanation in the summary.
3. Sovle --> solve
4. It might be good to show examples of data, or how the results are used if possible (I realize this is done plenty in the ApJS paper, but it also seems missing here).
5. I do not see a reference to the Magritte JOSS paper in the paper text.
6. References at end of sentences in full parentheses,  Matsumoto et al. (2023) -->  (Matsumoto et al. 2023)  and Danilovich et al. (2024) --> (Danilovich et al. 2024)
7. The last sentence of the statement of need could probably be simplified. Instead of talking about the forward/inverse problems, simpler language may be more clear in this context (e.g. fitting models to data). The ApJS paper also introduces these concepts very well, but so you could also point to that paper.
