### 2024-11-05 #7321 @s-goldman
@JBorrow @karllark I have reviewed the code and paper and found it to be worthy of publication with several minor changes.

This code is a significant scientific contribution, it is well written and documented, and conforms to coding standards and best practices. Much of the technical detail is outlined in [De Ceuster et al. 2024](https://ui.adsabs.harvard.edu/abs/2024arXiv240218525D/abstract), which pairs well with this paper. I have created several issues in the Pomme repository. Several of them are recommendations, but the primary issue with corrections for the paper is included in [Issue 8](https://github.com/Magritte-code/pomme/issues/8).
