### 2024-11-07 Magritte-code/pomme#8 @FredDeCeuster
Thank you for the feedback!
I've addressed the comments.
