### 2024-11-21 Magritte-code/pomme#8 @JBorrow
Thank you! It looks like @s-goldman's issues have also been resolved? If so I'll take one final look at the draft next week and recommend acceptance.
