### 2024-11-25 #7321 @JBorrow
Hi @FredDeCeuster, I'm looking to move this paper forward to the publication stage. TO do that, we will need a Zenodo deposit of the final version of the software with the following criteria:

- archive names (including order) and title in archive matches those specified in the paper
- archive uses the same license as the repo and is OSI approved as open source

We can then proceed to publication with some minor final checks.
