+++ joss/paper.md
@@ -39,7 +39,7 @@ bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+aas-doi: 10.3847/1538-4365/ad89a2
 aas-journal: Astrophysical Journal Supplement Series
 ---
 
