+++ src/pomme/__init__.py
@@ -1,5 +1,5 @@
 # Version number of this package (pomme)
-__version__ = "0.0.16"
+__version__ = "0.1.0"
 
 import torch
 torch.set_default_dtype(torch.float64)
