+++ conda/meta.yaml
@@ -1,6 +1,6 @@
 package:
   name: pomme
-  version: "0.0.17"
+  version: "0.1.0"
 
 about:
   home: "https://github.com/Magritte-code/pomme"
