### 2024-11-25 #7321 @JBorrow
Hi @FredDeCeuster you can use the arXiv paper which has a DOI for PyTorch: https://ui.adsabs.harvard.edu/abs/2019arXiv191201703P/abstract
