+++ joss/paper.bib
@@ -112,4 +112,5 @@ We demonstrate the efficiency of individual subsystems, as well as the overall s
 	publisher = {Curran Associates, Inc.},
 	author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
 	year = {2019},
+  doi = {10.48550/arXiv.1912.01703},
 }
