### 2024-11-25 #7321 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
