### 2024-11-26 #7321 @FredDeCeuster
Hi @JBorrow, do I have to do the check and accept?
Also, not sure if the XML files work... it seems like the check to merge is hanging...
