### 2024-11-26 #7321 @warrickball
Hi @FredDeCeuster — Though I agree it's not entirely clear, that message is intended for the relevant part of the editorial team (which in practice is me). I'll start checking the paper and metadata and probably suggest any editorial corrections via PR in the next few days.

Right now you don't need to do anything.
