### 2024-12-05 6b1b5c9
Editorial tweaks to JOSS paper
