+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @ARTICLE{DeCeuster2022,
        author = {{De Ceuster}, Frederik and {Ceulemans}, Thomas and {Srivastava}, Atulit and {Homan}, Ward and {Bolte}, Jan and {Yates}, Jeremy and {Decin}, Leen and {Boyle}, Peter and {Hetherington}, James},
-        title = "{3D Line Radiative Transfer \& Synthetic Observations with Magritte}",
+        title = {3D Line Radiative Transfer \& Synthetic Observations with {Magritte}},
       journal = {The Journal of Open Source Software},
      keywords = {radiative transfer, Python, astronomy, C++},
          year = 2022,
@@ -17,16 +17,18 @@
 @ARTICLE{DeCeuster2024,
        author = {{De Ceuster}, Frederik and {Ceulemans}, Thomas and {Decin}, Leen and {Danilovich}, Ta{\"\i}ssa and {Yates}, Jeremy},
         title = "{Bayesian model reconstruction based on spectral line observations}",
-      journal = {arXiv e-prints},
+      journal = {The Astrophysical Journal Supplement Series},
      keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2024,
-        month = feb,
-          eid = {arXiv:2402.18525},
-        pages = {arXiv:2402.18525},
-          doi = {10.48550/arXiv.2402.18525},
+        month = dec,
+       volume = {275},
+       number = {2},
+          eid = {44},
+        pages = {44},
+          doi = {10.3847/1538-4365/ad89a2},
 archivePrefix = {arXiv},
        eprint = {2402.18525},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240218525D},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024ApJS..275...44D},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -34,7 +36,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Matsumoto2023,
        author = {{Matsumoto}, Kosei and {Camps}, Peter and {Baes}, Maarten and {De Ceuster}, Frederik and {Wada}, Keiichi and {Nakagawa}, Takao and {Nagamine}, Kentaro},
         title = "{Self-consistent dust and non-LTE line radiative transfer with SKIRT}",
-      journal = {\aap},
+      journal = {Astronomy & Astrophysics},
      keywords = {radiative transfer, ISM: molecules, methods: numerical, infrared: ISM, Astrophysics - Astrophysics of Galaxies},
          year = 2023,
         month = oct,
@@ -53,7 +55,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Coenegrachts2023,
        author = {{Coenegrachts}, A. and {Danilovich}, T. and {De Ceuster}, F. and {Decin}, L.},
         title = "{The unusual 3D distribution of NaCl around the asymptotic giant branch star IK Tau}",
-      journal = {\aap},
+      journal = {Astronomy & Astrophysics},
      keywords = {stars: individual: IK Tau, stars: AGB and post-AGB, circumstellar matter, submillimeter: stars, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2023,
         month = oct,
@@ -100,7 +102,7 @@ archivePrefix = {arXiv},
 
 
 @inproceedings{Paszke2019,
-	title = {{PyTorch}: {An} {Imperative} {Style}, {High}-{Performance} {Deep} {Learning} {Library}},
[…]
