### 2024-12-05 #7321 @warrickball
@karllark, @s-goldman — As a matter of policy, note that this publication is part of a collaboration with AAS Journals, which means that AAS Journals make a small donation to the running costs of JOSS. You can find out more in the [original announcement](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing).
