### 2024-12-06 #7321 @FredDeCeuster
Thanks!
I’ve merged the PR.

> On 5 Dec 2024, at 22:15, Warrick Ball ***@***.***> wrote:
>
>
> Apologies for a slightly longer wait than I expected. I neglected to post here but I was waiting for the ApJS DOI to start resolving, which it now does. I've opened a PR <https://github.com/Magritte-code/pomme/pull/15> with what I hope will be the finishing touches. Once those are merged (presuming my changes are acceptable!) I'll do another test run of our publication workflow.
>
