### 2024-12-06 #7321 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
