### 2024-12-09 #7321 @warrickball
Hi @FredDeCeuster — I apologise but I noticed one final citation that needed to be tweaked so I've opened [another tiny PR](https://github.com/Magritte-code/pomme/pull/16). That'll be the last change!
