### 2024-12-10 #7321 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
